refactor(realtime): align durable replay ownership
