+++ apps/api-v1/src/create-rallar-server.ts
@@ -11,10 +11,8 @@ import {
   createRallarServerApplication,
   type RallarServerApplication,
 } from '@shared-server/rallar-facade/RallarServerApplication.ts';
-import {
-  RallarServerDataFacade,
-  RallarServerSystemFacade,
-} from '@shared-server/rallar-facade/RallarServer.ts';
+import { RallarServerDataFacade, RallarServerSystemFacade }
+  from '@shared-server/rallar-facade/RallarServer.ts';
 import { initRallarSystemWsTopics } from '@shared-server/rallar-system/ws-system-topics.ts';
 import {
   RallarRtcTopologyService,
@@ -25,15 +23,15 @@ import { GroupTopologyConfigRepository } from '@shared-server/rallar-system/repo
 import { GroupStateRepository } from '@shared-server/rallar-system/repositories/GroupStateRepository.ts';
 import { RtcRttRepository } from '@shared-server/rallar-system/repositories/RtcRttRepository.ts';
 import { RtcTopologySnapshotRepository } from '@shared-server/rallar-system/repositories/RtcTopologySnapshotRepository.ts';
-import { GroupTopologyManagementService } from '@shared-server/rallar-system/services/group-topology-management-service.ts';
-import { PSqlAdminOperationsStatsReader } from '@shared-server/postgres/admin-operations/PSqlAdminOperationsStatsReader.ts';
+import { GroupTopologyManagementService }
+  from '@shared-server/rallar-system/services/group-topology-management-service.ts';
+import { PSqlAdminOperationsStatsReader }
+  from '@shared-server/postgres/admin-operations/PSqlAdminOperationsStatsReader.ts';
 import { AdminSupportService } from '@shared-server/rallar-system/admin-support/AdminSupportService.ts';
 import { PSqlAdminSupportReader } from '@shared-server/postgres/admin-support/PSqlAdminSupportReader.ts';
 import { sendStateSyncMessage } from '@shared-server/rallar-system/state-sync-routing.ts';
-import {
-  readGroupGraphDiagnostic,
-  readScopedGlobalGraphDiagnostic,
-} from '@shared-graph/graph-diagnostics-service.ts';
+import { readGroupGraphDiagnostic, readScopedGlobalGraphDiagnostic }
+  from '@shared-graph/graph-diagnostics-service.ts';
 import type { RallarServerWsFacadeOptions } from '@shared-server/rallar-facade/ws-topic-router.ts';
 import type { Middleware } from './middleware-contract.ts';
 import { initialiseMiddleware, registerMiddlewareBackgroundTask } from './middleware.ts';
@@ -68,9 +66,9 @@ import { createRuntimeStateRepository } from './repository/createStateRepositori
 import { SpaStatisticsService } from '@shared-server/rallar-system/spa-statistics/SpaStatisticsService.ts';
 import { toWebRtcGroupKey } from '@shared/api/api-type-utils.ts';
 import type { GroupRef } from '@shared/api/group-types.ts';
-import {
-  createApiRtcTopologyAdminMetrics,
-} from './runtime/rtc-topology/rtc-topology-admin-metrics.ts';
+import { createApiRtcTopologyAdminMetrics }
+  from './runtime/rtc-topology/create-api-rtc-topology-admin-metrics.ts';
+import { readAdminClientIds } from './services/read-admin-client-ids.ts';
 
 export { RallarServerDataFacade, RallarServerSystemFacade };
 
@@ -397,11 +395,4 @@ export function createRallarServer(
     },
   });
   return rallarApplication;
-}
-
-function readAdminClientIds(): readonly string[] {
-  return (Deno.env.get('AUTH_ADMIN_CLIENT_IDS') ?? 'admin')
-    .split(',')
-    .map((value) => value.trim())
-    .filter((value) => value.length > 0);
 }
