+++ apps/api-v1/src/services/read-admin-client-ids.ts
@@ -1,0 +1,6 @@
+export function readAdminClientIds(): readonly string[] {
+  return (Deno.env.get('AUTH_ADMIN_CLIENT_IDS') ?? 'admin')
+    .split(',')
+    .map((value) => value.trim())
+    .filter((value) => value.length > 0);
+}
