+++ packages/shared-server/rallar-system/topology/replay/rtc-topology-replay-decision.ts
@@ -101,7 +101,7 @@ function readPublicationSnapshot(
   entry: RtcTopologyDeliveryLogEntry,
 ): RallarOverlayTopologySnapshot {
   try {
-    const snapshot: unknown = JSON.parse(publication.message.payload.resource);
+    const snapshot = JSON.parse(publication.message.payload.resource);
     validateTopologySnapshot(snapshot, entry.groupRef);
     return snapshot;
   } catch {
