+++ packages/tests/shared-server/mutation-routing-markers.ts
@@ -34,7 +34,9 @@ export const ROUTING_SOURCE_MARKERS = {
     'enqueueAuthorisedWsClientConnect',
   ],
   'packages/shared-server/crdt/RallarCrdtServer.ts': ['mutationIngress.enqueueUpdate'],
-  'packages/shared-server/rallar-system/ws-system-topics.ts': ['enqueueRtcRttMutation'],
+  'packages/shared-server/rallar-system/topology/rtt/init-rtc-rtt-topic.ts': [
+    'enqueueRtcRttMutation',
+  ],
   'packages/shared-server/rallar-system/services/ws-lifecycle-service.ts': [
     'enqueueClientSessionDisconnect',
     'enqueueGroupSessionCleanup',
